feat: robot.txt 추가, 메타데이터 수정하며 컨테이너 제거
